fix: use double for Polygon volume DTO to handle float responses
Polygon API occasionally returns volume as a float (e.g. 1234567.0),
which fails to deserialize into Int64. Changed the DTO field to double
and cast to long at the mapping boundary.

+++ src/Signavex.Infrastructure/Polygon/PolygonMarketDataProvider.cs
@@ -54,7 +54,7 @@ public class PolygonMarketDataProvider : IMarketDataProvider
                 r.High,
                 r.Low,
                 r.Close,
-                r.Volume
+                (long)r.Volume
             ));
         }
         catch (Exception ex)
